refactor(cli): home parse tests and bind-gate tests
Move clap parse coverage out of main.rs into cli_tests.rs, and move the
env-mutating bind-gate tests next to ensure_bind_allowed in serve.rs.
main.rs is now the dispatcher (~470 lines) instead of a 1750-line dump.
